Recreate package-lock.json
Even without any changes to package.json,
deleting and recreating package-lock.json from scratch
changes versions of a few dependencies.
Furthermore, it reduces cases where we previously included the same
library several times in different versions and replaces it with a
single version. Not sure why that even happened.
Due to this effect, the bundle size is even a little bit smaller now.