Do not use directory hierarchy for work dir of overlayfs
Each overlayfs instance needs a "work" directory
that should not be used for anything else.
So far, if we mounted two overlayfs instance for e.g. "/foo" and
"/foo/bar", we would use ".../work/foo" and ".../work/foo/bar"
as work directories.
However, the latter directory is actually inside the former directory
and could create problems for the first overlayfs instance.
This also creates a warning in the kernel log.
So we avoid this by using a flat directory list like
".../work/1" and ".../work/2".

The same problem also exists for the "upper" directory,
but there is is more difficult to solve
(and it is unclear whether it is really necessary).