Fix error message if command-line arguments use invalid unit
We did provide a nice message already internally,
but the library for parsing command-line arguments did not pass it to
the user. Now we make sure that this happens.

So far these parsing methods raised ValueError,
but argparse requires an ArgumentTypeError,
otherwise the exception message is thrown away.
However, some other callers except these methods to raise ValueError.
The easiest way to solve this is to raise something
that is both a ValueError and an ArgumentTypeError.