Improved error handling and message
Standard style in BenchExec is to pass a fatal error message
to sys.exit() instead of separating it out into a log message.
This keeps the termination reason attached to the exception that will
propagate upwards.
Similarly, a single log message should not be split into two logging
statements anyway.

Also improve the error message to tell the user what this is about.