Color category names in filter definitions of HTML tables
Each result category is associated with a certain color.
We can also use this color for the actual category names
in the filter box to make the connection clear to users,
i.e., "correct" will be shown in green etc.
Chromium even allows us to apply these colors to the filter drop-down.