Encode environment variables with illegal characters in result XML
It is not possible to use all kinds of characters in XML 1.0,
some like the ASCII control characters are forbidden even if escaped
like "&#x1b;". But such characters can occur in environment variables.
So we need to put our own encoding on top.
We use Base64 for this and mark it in the XML with a new attribute.

This is done only if problematic characters actually appear in a
variable, in which case BenchExec would crash previously.
So nothing changes for existing users of BenchExec.
Furthermore, BenchExec itself does not actually use these tags (yet),
and we also have no known other users of them.

Fixes #828.