Remove special casing for cgroup versions from RunExecutor
The cgroups classes should abstract away from the differences,
including faking a controller like "freezer" if it is no longer
a separate controller on v2.
Furthermore, we cannot just assume that these features are always
usable on v2 systems because it can still happen that we do not have a
usable cgroup at all.