Rewrite main logic for initializing cgroups
There were at least two problems:
- Deeply nested cgroups created in tests due to move_to_scope() there.
- runexec not working without manual cgroup creation like benchexec.

We need to have some global initialization logic
that works for all circumstances and which is safe to be called more
than once without creating lots of deeply nested cgroups
in order to solve this.
It even needs to be thread safe for concurrent callers of RunExecutor.

What is missing in this version is proper error handling:
For example, if we cannot move the process to a new systemd scope,
we do not want to produce an error immediately because cgroups might
actually be optional. So we need to produce the error only later if we
find out that we actually require cgroups, but then the information
about why we cannot have cgroups is lost.
So for now we just get a generic error message and not one showing the
reason why cgroups are unusable.
This will have to be implemented.