Remove restriction to well-known controllers for cgroups v2
This would crash if somebody calls us in a cgroup with controllers
enabled that we do not know.
For cgroups v1 we have this check because unknown controllers
can do weird things and it is not in general safe or even possible
to ignore them.
However, controllers for cgroup v2 are expected to be more well behaved.