Refactor CgroupsV2.create_fresh_child_cgroup()
The handling of move_to_child=True is needed only in one specific case,
so we can move it there and simplify the method.
Furthermore, we can simplify the handling of delegating controllers,
because if a cgroup has processes, we know it has no controllers.
So we do not need to remove controllers before creating the child.