Fix output of pressure stall time with cgroups v2
Because of the use of type float,
for small values (which can occur here)
the output would be in scientific notation,
but we promise decimal notation.
Furthermore, float can of course be imprecise.
So we switch to the type Decimal and make sure it is printed properly.

It would be good to also switch to Decimal for other time values,
but we should make that change separately.