First step at better error handling for cgroups v2
As discussed in 4ef31426, we currently do not get error messages that
explain why cgroups are not working (e.g., missing pystemd or something
like this).
The current error messages for all the different situations are only for
cgroupsv1.
This commit moves the existing error messages to cgroupsv1.py
and adds new error messages for situations that can exist with
cgroupsv2.
It is likely that we need to add more cases or better explanations.