While terminating all processes, unfreeze all child cgroups
If a process in a run creates a child cgroup with processes
and freezes that cgroup, we would otherwise fail to perform
proper cleanup of all processes.
This is not possible in the default configuration,
but if the user gives access to cgroups to the processes in the run,
e.g., in order to be able to nest runexec.

Fixes #840 as least as far as possible,
i.e., if BenchExec has access to and uses the freezer cgroup.
In principle, it might be possible that there is something that prevents
BenchExec from using the freezer cgroup (it is optional in container
mode) but still allows the inner processes to use it,
and then we would still fail.
But I cannot imagine such a situation except if some external process
interferes (in which case we cannot guarantee anything anyway).

This commit also does not change behavior of containerexec,
it just fixes runexec and benchexec.