Guard against child cgroups with unexpected permissions
In the case where the run's processes have cgroup permissions,
they might create child cgroups and then run "chmod 000" on them
or on the files contained within.
This would crash runexec during process termination
and prevent proper cleanup.

Of course, if the child cgroups are changed such that they do not longer
belong to our user account, all hope is lost.
But the processes inside the run cannot do this.