Fix mount handling in case a mount point does not exist
This is a generalization of ce48ec1, which solved the problem for mount
points (in the kernel's list) that are inaccessible,
i.e., inside a directory where we do not have access.
But there can also be mount points in the kernel's list that simply do
not exist, if we are running inside a container where the container
runtime has mounted an empty directory over the parent of a mount point
(just like we do, but we actually prevent this case since 82c035).

This makes BenchExec work inside Podman containers
without having to pass "--hidden-dir /sys".