Add a cache for working_directory() of containerized tool-info module
The goal is to avoid the multiprocessing overhead for this method,
which is called often but always returns the same result.
The multiprocessing overhead is typically not large,
but in some situations it is:
https://github.com/python/cpython/issues/98493

On systems that are affect by this, it can save several minutes of
execution time if there are 10000 runs.