Improve handling of duplicate results in table-generator
If a run set contains more than one result for the same task,
we keep only the first result.
This commit improves how this is achieved.
Previously, duplicate results were filtered during the merge step,
now we do this immediately after creating the RunSetResult objects
in the collect_data step.

One advantage is that a warning about duplicates is now printed
even if --common is used, this was previously missing.

Furthermore, the data structure is now easier to understand.

As a side effect, the order of the warnings for duplicate results
changes now.