Optimize common case of the "merging results" step in table-generator
The worst case complexity of this method is O(n*n*m)
where n is the number of tasks in the table and m the number of run sets.
This stays the same, but for common cases we can improve it to O(n*m),
which makes a huge difference because m is typically small compared to n.

Such a common case where the new performance is better
is if all m run sets contain the exact same set of n tasks,
which likely is even *the* most common case.

A case that was already fast before and stays the same
is if all m run sets contain (mostly) disjunct task sets,
but this is rare.

The worst case now is if all but the first run set
contain an alternating sequence of tasks [e1, n1, e2, n2, ...]
where all e tasks already exist in the previous run sets
and all n tasks are new in this run set.
But this should be extremely rare.