Drop support for Python 3.6
This was announced already in July 2022 for version 3.12 of BenchExec,
so there was plenty of time for getting aware of it.
We postponed it because there was no urgent need,
but now we want to start using a feature of Python 3.7
(cf. 1b5010f3e7fd7da413ce2b83b7ef03e349f86124).