Hide message about missing properties in table-generator
There are many use cases where properties are not used,
but table-generator would log one line about every row
with a missing property just because it cannot compute
the number of expected true/false tasks in this case.
But this disturbs the output of all users who do not use properties
a lot, especially because this log line comes in the very end
and hides the (useful) log lines with the names of the output files.

While not explicitly discussed there,
this can be seen as part of #898.