Require that we are the only process in our cgroup for cgroupsv2
So far we accepted other processes from the same process group,
but this would mean that we move these other processes into a different
cgroup, and maybe we should not do this.
Let's wait for whether some use case appears where it is relevant
that we do not require to be the only process in the cgroup.