More debug output if something happens during tool start
We have a general "except OSError" block here,
which catches exceptions not only from the actual tool start,
but also from lots of cleanup code.
This can be confusing.
In general it would be better to clean this up properly
and handle all such errors specifically,
but as a first quick improvement we can add the stack trace
to the debug log if this happens.
Then we can at least see where the problem was.

This can be triggered for example with
bin/runexec --debug --full-access-dir / rm output.log
because this crashes runexec when it tries to access the output log.