Fix a crash in cgroup cleanup in case of nested cgroups
Sometimes a nested cgroup vanishes while we are performing cleanup,
and this would lead to a cleanup.
I had thought that this should never happen because we ensure
that all processes are already killed,
but maybe there is some delayed cleanup in the kernel or so.
So we guard against this.