Frequency shown for CPUs should be maximum allowed frequency
CPUs have a technical frequency limit,
but system administrators can also configure a lower limit.
The value that is relevant for benchmarking users
is the actual limit, not the theoretical one
(this one could also be inferred from the CPU name and the datasheet).
So we read the latter (actual) limit.

Fixes #918.