Add --cgroup-access to containerexec for cgroupsv2
The cgroup namespaces from the previous commit provide better isolation,
so using them also makes sense for containerexec, not just for runexec.
However, containerexec currently does not use and require cgroups at
all, and we do not want to make them mandatory,
such that containerexec stays usable for users without cgroup access.
So we add a new argument --cgroup-access to containerexec
that triggers use of cgroups for the run (without any limits etc.)
and uses cgroup namespaces to provide a usable cgroup for the process
inside the container.
We support this flag only on systems with cgroupsv2
because with cgroupsv1 the cgroup namespaces do not work as well.

runexec and benchexec do not get this argument,
because they use cgroups anyway and thus --cgroup-access is implied.

This closes #436.