Make failure of sethostname in container only a warning
Podman containers forbid sethostname via seccomp
even though it would be safe.
To use BenchExec inside these containers, users currently have to pass
--keep-system-config, but this disables other desirable configs as well.
If we cannot change the hostname, this is not crucial,
it will only mean that the benchmarked tool will see the real hostname,
but as we are likely in some other container, it will the that
container's hostname.
So there is no need to fail hard.