Let check_cgroups produce better error message for missing CPUSET
There is this kernel bug for cgroups v2 that prevents us from using
CPUSET but everything else is working.
In this case we provide a nice error message with a workaround (6cbd9fc5),
but check_cgroups did not do this so far because it terminated itself
too early. Now we continue and let RunExecutor print the message.