Silence warning about missing CPUSET if not required
So far, BenchExec prints a warning about every missing cgroup
controller, and later on an error message if that controller is strictly
required. But CPUSET is only required for core limits, and has no other
use. This is different from e.g. MEMORY which is required for memory
limits but also provides memory measurements.
So lets silence the warning about missing CPUSET and keep only the error
message.

check_cgroups is not affected because if forces CPUSET to be required.