Handle case of missing cgroups delegation for systemd and cgroups v2
The default configuration at least on Ubuntu is that users only have
access to the memory and pids controllers, not to all of them,
because the system-wide systemd instance does not delegate the other
controllers to the user-specific systemd instance.
I had not noticed this earlier in my testing
because long ago I had already configured systemd to delegate
everything.
Without the delegation of other controllers only a limited feature set
of BenchExec is available.

This means that installation of BenchExec still requires some
system-wide configuration, even with cgroups v2.
So we update the documentation accordingly,
provide an appropriate error message for users,
and extend the Debian package to perform this automatically.