Allow get_value_from_output in tool-info modules to return numbers
If a tool-info module wants to return numbers for additional columns
after doing some calculations, it is more convenient to be able to
return the numeric instances directly.
Furthermore, especially for Decimal we can do better than the standard
conversion with str().

This adds support for numeric return types to this method,
both when used inside benchexec and when used inside table-generator.

The tool-info module of ESBMC can benefit from this,
so let's do so.