Initial set of config for GitHub actions with Black, Reuse, and Ruff
Unfortunately, GitLab CI still does not work for GitHub PRs from forks,
and Travis has become unusable.
So reviewing PRs from forks is tedious
because these checks have to be executed manually.

We could use AppVeyor, which we already use for tests on Windows,
but GitHub Actions are more common and likely the UI is better.
So let's try this.