Avoid uses of defaultdict in resources.py
Uses of plain dicts may catch errors in callers earlier.
Furthermore, some of the functions even returned a defaultdict
in some cases and a plain dict in other cases.
The return type should be consistent.

With dict.setdefault() the use of a plain dict
is almost as convenient as a defaultdict.