Restore tests for core allocation from main branch
Because the tests were changed significantly together with the
implementation of the core allocation in this branch,
it is hard to see whether the tests still test the same things.
But we want regression tests that ensure that the new allocation
behaves in the same way.
So we restore the old tests and keep the new tests in a separate file
until we are sure that there are no regressions,
and then we can unify the test suites.

For now the old tests fail due to API changes,
this will be adapted next.