Restructure reading of siblings information
So far we read the information about the hyperthreading hierarchy level
differently from the other levels.
This made the code more difficult to understand,
and the way how the ids in the hierarchy_levels[0] dict were chosen
differed from the other levels.
But we can also read this information in the same way as for the other
levels, so let's do this.

We still also need to use the previous way of reading all siblings from
a given list of cores, but we can also simplify that
and the separation of concerns still provides
an understandability benefit.