Support arbitrary many cache hierarchy levels for core allocation
The allocation algorithm already supports an arbitrary number of levels,
so we can future proof the allocation
and read all information about cache levels
that the kernel provides.

We can also use the assumption that caches are named the same
across all cores, and read the cache names only once
instead of separately for every core.