Handle proc fs as usual if mounted somewhere else than /proc
For /proc, we need directory mode "read only",
but we only checked for fstype being "proc" and not for the path.
This meant that BenchExec would mount proc mounts
in other places than /proc as read only and ignore the directory mode
specified by the user.
This creates problems if the parent directory is hidden, for example.

Fixes #988