Polish documentation for using BenchExec in a container
We want to provide a full usable command line
in the main installation docs and not have the information
split on two documentation pages.
Using the container mode of BenchExec is not so rare anymore
that we document it separately.

And because people misunderstand what to do for Docker,
also copy the instructions for Docker into the main documentation.