Handle invalid cgroup setup more gracefully for cgroups v2
Inside a container with an invalid cgroups setup,
such as cgroupns but using the host's cgroup filesystem,
it can happen that our cgroup is shown by the kernel as "/../foo".
This is an unusable cgroup.
So far BenchExec would crash due to this,
now we at least handle it like a cgroup without permissions.

Fixes #1047.