Add a fallback logic for locating a usable cgroup for cgroups v2
For cgroups v1 we already use /system.slice/benchexec-cgroup.service
as a fallback cgroup if it exists and we cannot use our current cgroup.
For cgroups v2 we do not yet have such a fallback.
However, implementing this makes it much easier for people
on systems without systemd (e.g., in a container)
to setup and configure cgroups for BenchExec, so let's do this.

Background discussion is in #1048,
from which also the suggested commands in the documentation are taken.

Thanks to @incaseoftrouble for reviewing the design of this.