Fix crash in fallback logic if /benchexec cgroup does not exist
It is not nice to have to double-check this here
but CgroupsV2.from_system() is also used for other cases
and we cannot easily change it to return only a dummy cgroup
if the cgroup does not exist.
Once we get rid of cgroups v1 we might be able to refactor
and clean up a lot of this.