Fix stdin handling when getting version number from benchmarked tool
In the function that most tool-info modules use
for getting the version number from the tool,
we did not specify what stdin of the tool should be.
The unfortunate default for this is the same as stdin
of the calling tool, i.e., BenchExec.
This means that if a tool would read from stdin
while we call it to get its version,
it would either hang due to no input
or even read from what the user attempted to pass to BenchExec.
Both is not good and this happened in #1045.

Now we explicitly specify /dev/null as stdin,
such that when attempting to read the tool just gets EOF.
We also use /dev/null for stdin by default for the actual tool execution
since 564f18f (BenchExec 0.5), so I do not expect any problems.
Tools who do not read from stdin are not affected,
and for tools who do read it is just an improvement
that they get EOF instead of waiting for input.