Fix benchexec when running in empty cgroup with cgroups v2
We document that one can start BenchExec in an otherwise cgroup.
However, this works only for runexec so far,
because we check that the current cgroup has exactly one process.
But benchexec starts the tool-info module in a separate process
if container mode is used, and so this check would fail.

We need to relax this and also allow child processes of us
in the given cgroup, and then later on move these child processes
into a different cgroup like our own process
(the existing code does this already).