Use groupadd instead of addgroup in our .deb-package installation script
addgroup is a user-friendly high-level command,
whereas groupadd is a low-level command.
I noticed out that not in all containers addgroup is available,
and then the installation script reports an error.
AFAIS there should be no difference between
"addgroup --system" and "groupadd --force --system",
so we can switch to the latter, which exists everywhere.