Adapt tool-info module to new option style of CPAchecker
CPAchecker will soon use command-line arguments with "--" as prefix.
The old arguments will continue to be supported,
but we want to use the new ones in the tool-info module when possible
in order to not confuse users.
So let's let the tool-info module adapt to the options
that the user specified in the benchmark definition.

Cf. https://gitlab.com/sosy-lab/software/cpachecker/-/issues/1218