Fix tool-info module of CPA-witness2test: executable is required
Previously, the tool-info module of CPAchecker ensured
that the executable of CPA-witness2test is covered by program_files(),
because the scripts/ directory is part of the program files.
But for new versions of CPAchecker we do not add all of bin/
to the program files, so we need to explicitly add cpa-witness2test.