Use compatibility level 12 for building Debian packages
According to the man page debhelper-compat-upgrade-checklist
level 11 is deprecated and not recommended.
A test build on level 12 produces an identical .deb package (same hash),
so it seems we can safely upgrade.

Also switch how we specify the compatibility level:
In debian/control instead of debian/compat,
because right now we need to specify it in two places.