changed code under test: added additional assertions, copying context from default context
The current implementation fails under pytest, as the context (as given by getcontext())
is already derived from the DefaultContext *before* we change the rounding to our
preferred value. Thus, the assertion correctly fails, as the first process still has
the default ROUND_HALF_EVEN value, even though all decending processes do possess the
correct value of ROUND_HALF_UP.

As a solution, we can copy the DefaultContext as the context of the initial thread.
As the assertion afterwards becomes very obviously true, additional assertions are added,
to check if the individual processes have the correct rounding.