Fix crash in CPAchecker's tool-info module
We have a heuristic that checks if CPAchecker is used from source
but misses a rebuild. This would crash if broken symlinks or other
file-access problems occur.
As it is just an optional feature, we can ignore such problems
and print a warning.

An example where this happened was for example
system-wide installation of CPAchecker
with the executable being /usr/bin/cpachecker
and /usr/src existing.