Remove outdated version pins from release.sh
Most of them are satisfied anyway and/or irrelevant
because build will happen in an isolated environment,
and we declare build dependencies in pyproject.toml.
Specifying them here was necessary in the past.

We specifically do still need setuptools<58 for executing the tests.
Document this such that we know when we can remove this.