Refactoring: allow to pass back a cgroups instance from executor
Currently RunExecutor creates a cgroup instance and uses that.
But sometimes the underlying executor needs to create a nested cgroup
and put the tool into that. Now we pass it back to RunExecutor
such that it can make use of it for the measurements.
The effect is that we can use different cgroups for limits
and for measurements (and killing processes).

But in this commit we still pass only the original cgroups instance back,
so there is no behavior change.