Improve integration of LXCFS and virtualize /proc/cpuinfo
We recommend to install LXCFS together with BenchExec,
because we use that to virtualize for example /proc/uptime in the container.
However, a main use case of LXCFS is to virtualize files
that contain information about the system such as the available CPU
cores in /proc/cpuinfo.
We never advertised this, but I assumed this was working all the time.
I found out that it never worked, though.

The reason is that LXCFS is using the limits configured for the init
process of the container, but our init process has no limits,
it is not part of the same cgroup as the other processes in the container
(on purpose, because we do not want to measure its resource consumption).
So now we create yet another cgroup for the init process
that is below the one with the limits
but outside of the one that is used for measurements.
Note: A single runexec execution will now create up to 5 cgroups.

This is made possible due to the separation between the cgroups
for limits and for measurements in the last commits.

With this change, /proc/cpuinfo now shows only the cores available in the
container if LXCFS is running.
This helps processes in the container to see how many CPU cores
they are allowed to use and for example to decide how many threads to spawn.

Fixes #1070