Also virtualize /sys/devices/system/cpu with LXCFS if available
Like for /proc, LXCFS provides a virtualized /sys/devices/system/cpu
that only shows the allowed cores.
Of course we want to mount that in the container as well,
at least if the user has not requested /sys to be hidden
or have full access to the host directory.

Fixes #1069