Cleanup old reference to .coveragerc left over from 528a3ad0
We need to pass the value of the config file for coverage,
and we moved from .coveragerc to pyproject.toml in that commit,
but forgot to update the CI job.
It continued to work because coverage hard-codes the value
".coveragerc" and uses the standard search for config files
instead of requiring this specific file,
but using the name of the actual config file is still cleaner
and less confusing.