Call poll() after the process finished to update returncode in Popen object
Popen in the subprocess module holds the exit code of the process in its internal state,
which is only set if the provided methods to wait for process termination are used
or poll is called. If the exit code is not set, a bogus warning message occours on
destruction of the object.