Change internal paths used for fuse-overlayfs mounts
The temp_base directory (.../temp) is the one that BenchExec uses
to store output files of the tool, and after a run we iterate through it
and copy files from there to the output directory.
Thus we should not use it for internal stuff.
But the work_base directory is fine for that.
So let's move the fuse mountpoint to work_base as well.