added more flexible way to to generate layer data for tests
To allow easier generation of new tests (where we ideally can automatically generate tests for a large number
of (also weird) CPU configurations), it's desireable to be able to specify arbitrary layers more or less directly, without
having to create new test classes.

The final goal is to be able to generate machine configurations given a single argument (or two arguments: layer configuration as a list
and total core count), so we can utilize pytest to write easily maintainable test cases.