Skip cgroupsv1 tests on intermediate Python versions for CI performance
It should be sufficient to run these tests only on the oldest and the
newest supported Python version, because it is unlikely that something
works for these two but fails for a version in the middle.
And we still run the tests for all Python versions on a cgroupsv2
system.