Skip tablegenerator tests in CI in combination with cgroupsv1
table-generator does not depend on system details like the cgroups
version, so it is enough to run its tests once in CI.
We do so in the main test runs (for cgroupsv2),
so we can skip them in the test runs for cgroupsv1.
This should speed up each of these test runs by ca. 1 minute.