Fix runs marked as failed on OOM for cgroupsv2
Because in 1049c39 we changed the cgroup hierarchy such that the memory
limit (deliberately) also applies to the init process of our container,
the kernel would kill our own child process together with the
benchmarked process together on OOM.
While the current commit does not guarantee that this is avoided
(there seems to be no solution for this that allows LXCFS to be used),
it should fix this in practice (depending on how well the OOM killer
chooses its targets).

This fixes #1098.