Improve cgroups cleanup and fix warning about non-removable cgroup
In the past, we had to take special care to leave one particular cgroup
present when we remove cgroups, because we needed it for reading a value
later.
However, since the restructuring of bcf45157 and cd2c6a8a
we actually do not even attempt to remove that cgroup anymore before
reading the memory anymore, because we first only make sure that the
benchmarked process is killed and any cgroups created by it are removed,
then read the measurements, and then perform the full cleanup of the
cgroup hierarchy.
But because we still kept the sub-cgroup during the full cleanup,
we got a warning that we cannot remove the parent cgroup.

So we can remove this special case, and this also gets rid of the warning.