Ultimate: fix search for available java
The previous search failed to use the java version on the PATH,
as glob.glob("java") always returned the empty list.
The new version coincides more closely with the implementation
in the Ultimate.py wrapper script, and thus should ensure that
the same java version is used from the tool info module and by
Ultimate itself.

Also not explicitly that the two implementations should match.