preparations to remove fixed xxxCoresPerRun methods
To be able to remove the fixed functions (and replace them with one which takes arbitrary numbers of cores), all calls to
the existing methods are redirected to a new method. This removes all direct references to mainAssertValid, allowing
us to refactor the code without having to modify each test at the same time.