add decorator to dynamically add test cases for any amount of CPUs used
Currently, all tests are limited to 1-4, 8 CPUs used with adding test cases taking a bit of work.
This decorator dynamically adds a test case to test if the actual assignment matches the expected
assignment to a test class. It doesn't use the fixed versions of xxxCoresPerRun, which allows us to
remove them later entirely.