added assertion for new machine definitions
The added assertion is required to ensure the new definitions (which allow arbitrary CPU layouts¹) are actually
identical to the existing ones. It will be removed together with the old definitions after the refactoring

Note that the definition is based on the physical CPU layout, with the number of virtual cores per core being the second value
of the (<cpu_layout>, <virtual_cores_per_core>) tupel.

¹) with the only restriction that we require CPUs in multi CPU systems to be identical, i.e. the layout needs to be symmetric