added machine definitions based on new format
As noted in the previous commit, those definitions are based on the physical layout of the CPUs, with
the number of hyperthreading / virtual cores per core being a seperate argument. This has the upside that the machine definitions do
not need to be internally consistent, i.e. we do not need to take care if the total number of cores is actually possible
with the number of virtual cores.